updating dotmanager.pl

+++ dotmanager.pl
@@ -30,45 +30,56 @@ my %files = (
     ".zshrc"        => $root,
 
     #scripts
-    ".lock.sh"      => $root,
-    ".mysuspend.sh" => $root,
-
+    ".lock.sh"      => $root . "/.bin",
+    ".mysuspend.sh" => $root . "/.bin",
+    ".urxvt-colors" => $root . "/.bin",
+    
     #terminal-colors
     "my-default.sh" => $root . "/.dynamic-colors/colorschemes",
     "gummy.sh"      => $root . "/.dynamic-colors/colorschemes"
 );
 
 =doc
+    Reads input prompt and returns answer (taken from StackOverflow)
+=cut
+
+sub prompt {
+    my ($query) = @_;
+    local $| = 1;
+    print $query;
+    chomp(my $answer = <STDIN>);
+    return $answer;
+}
+
+=doc
     Install dotfiles of the current repository for the current user in the system
 =cut
 
-sub copy {
-    my ($source_file, $destiny_file) = @_;
-    
-    if (-d $source_file) {
-        if (!dircopy($source_file, $destiny_file)) {
-            return 0;
-        }
-    } else {
-        if (!fcopy($source_file, $destiny_file)) {
-            return 0;
-        }
-    }
-    return 1;
-}
-
 sub install {
+    my $skip = shift;
+    
+    print("Starting the Installing Process \n");
+
     foreach my $key (keys %files) {
         my $source_file = $current_dir . "/" . $key;
         my $destiny_file = $files{$key} . "/" . $key;
 
-        print("Installing " . $key . " in " . $destiny_file . "\n");
+        if (not $skip) {        
+            my $answer = prompt("Are you sure you want to install $key ? [y/n]");
 
+            if (not ($answer eq 'y' || $answer eq 'Y')) {
+                print("Skipping $key update\n");
+                continue;
+            }
+        }
+        
+        print("Installing " . $key . " in " . $destiny_file . "\n");
+        
         my $result = `cp -TR $source_file $destiny_file`;
         
         if ($result) {
             say "Failed to install " . $key . " is probably being used by other application or it does not exists";
-        }
+        }    
     }
 }
 
@@ -77,28 +88,50 @@ sub install {
 =cut
 
 sub update {
+    my $skip = shift;
+    print("Starting the Uploading Process\n");
+
     foreach my $key (keys %files) {
         my $source_file = $files{$key} . "/" . $key;
         my $destiny_file = $current_dir . "/" . $key;
 
+        if (not $skip) {            
+            my $answer = prompt("Are you sure you want to upload $key ? [y/n]");
+
+            if (not ($answer eq 'y' || $answer eq 'Y')) {
+                print("Skipping $key update\n");
+                continue;
+            }
+        }
         print("Uploading " . $source_file . " in " . $destiny_file . "\n");
-
+        
         my $result = `cp -TR $source_file $destiny_file`;
-
+        
         if ($result) {
-            say "Failed to update " . $key . " is probably being used by other application  or it does not exists";
+            print("Failed to update " . $key . " is probably being used by other application  or it does not exists\n");
         }
     }
 }
 
-if ($#ARGV != 0) {
-    print("Usage: perl dotmanager.pl [OPTIONS]\n --install [Install current files in repository to system]\n --update [Update files in repository with ones from the system]\n");
+my $usage = "Usage: perl dotmanager.pl [OPTION] [BYPASS]\n
+  Options: 
+    --install [Install current files in repository to system]\n
+    --update [Update files in repository with ones from the system]\n
+  Bypass:
+    --skip [Skip Confirmations]\n
+    ";
+    
+if ($#ARGV == 0 || $#ARGV > 2) {
+    print($usage);
 } else {
-    if ($ARGV[0] eq "--install") {
-        install;
-    } elsif ($ARGV[0] eq "--update") {
-        update;
+    my $option = $ARGV[0];
+    my $bypass = ($ARGV[1] and ($ARGV[1] eq "--skip")) or 0;
+    
+    if ($option eq "--install") {
+        install($bypass);
+    } elsif ($option eq "--update") {
+        update($bypass);
     } else {
-        print("Invalid argument");
+        print("Invalid argument\n$usage");
     }
 }
